organize project in packages
