+++ src/compiler/visitors/semantics_check/formatter.py
@@ -1,15 +1,21 @@
-import compiler.visitors.visitor as visitor
-from ..cmp.ast import ProgramNode, ClassDeclarationNode, FuncDeclarationNode
-from ..cmp.ast import AttrDeclarationNode, AssignNode, CallNode
-from ..cmp.ast import (
-    CaseNode,
+from ...cmp.ast import (
+    AssignNode,
+    AtomicNode,
+    AttrDeclarationNode,
+    BinaryNode,
     BlockNode,
-    LoopNode,
+    CallNode,
+    CaseNode,
+    ClassDeclarationNode,
     ConditionalNode,
-    LetNode,
+    FuncDeclarationNode,
     InstantiateNode,
+    LetNode,
+    LoopNode,
+    ProgramNode,
+    UnaryNode,
 )
-from ..cmp.ast import UnaryNode, BinaryNode, AtomicNode
+import compiler.visitors.visitor as visitor
 
 
 class FormatVisitor(object):
